refactor(formatter): split word formatter by domain
